refactor: se añadieron las secciones faltantes del dashboard de medico
