+++ frontend/src/components/sections/Medico/DashboardMedico/DashboardMedico.tsx
@@ -1,6 +1,6 @@
 import { TableContextProvider } from "@context/table.provider";
 import { DashboardMedicoContent } from "./DashboardMedicoContent";
-import { MEDIC_LINKS } from "./MEDIC_LINKS";
+import { MEDIC_LINKS } from "../MEDIC_LINKS";
 import { SectionWrapper } from "@components/layout/SectionWrapper";
 import { getPacientes } from "@services/getPacientes";
