+++ frontend/src/components/sections/Medico/GestionarAgenda/GestionarAgenda.tsx
@@ -1,0 +1,10 @@
+import { SectionWrapper } from "@components/layout/SectionWrapper";
+import { MEDIC_LINKS } from "../MEDIC_LINKS";
+
+export const GestionarAgenda = () => {
+    return (
+        <SectionWrapper asideBarItems={MEDIC_LINKS}>
+            <></>
+        </SectionWrapper>
+    );
+}
