+++ frontend/src/components/sections/Medico/HistorialMedico/HistorialMedico.tsx
@@ -1,0 +1,10 @@
+import { SectionWrapper } from "@components/layout/SectionWrapper";
+import { MEDIC_LINKS } from "../MEDIC_LINKS";
+
+export const HistorialMedico = () => {
+    return (
+        <SectionWrapper asideBarItems={MEDIC_LINKS}>
+            <></>
+        </SectionWrapper>
+    );
+}
