+++ frontend/src/components/sections/Medico/MEDIC_LINKS.tsx
